Fix training of TabM with share_training_batches=False.
When the `share_training_batches` option was added to `bin.model.Model`
in [this commmit](https://github.com/yandex-research/tabm/commit/e284de3d1b592aa621365d684372f076baedc014),
it turned out that the training code in `bin/model.py` was not working
correctly when `share_training_batches` is set to `False`.

**Importantly**, the results in `exp` were obtained *without* the bug,
so they are correct.